requirements.txt
# Core RAG Framework
langchain==0.1.0
langchain-community==0.0.13
tiktoken>=0.12.0

# LLM Integration
ollama==0.6.0
# Embeddings & Vector Store
sentence-transformers==2.3.1
chromadb==0.4.22
faiss-cpu==1.7.4

# Sparse Retrieval
rank-bm25==0.2.2

# Reranking
transformers==4.36.2
torch==2.1.2

# Document Processing
pypdf==3.17.4
python-docx==1.1.0
beautifulsoup4==4.12.3
lxml==5.1.0
arxiv==2.1.0

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.12.5
pydantic-settings==2.1.0

# Evaluation
ragas==0.1.1
datasets==2.16.1
nltk==3.8.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
numpy==1.26.3
pandas==2.1.4

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.28.1

# Caching (optional)
redis==5.0.1

# Jupyter for notebooks
jupyter==1.0.0
ipykernel==6.29.0
matplotlib==3.8.2
seaborn==0.13.1
